requirements.txt
flask
pillow
PyPDF2
fpdf